@@ -106,9 +107,8 @@
     "lodash": "^4.17.21",
     "mathjax": "^3.2.2",
     "nanoid": "^5.0.8",
-    "next": "^14.2.23",
+    "next": "^15.3.0",
     "next-cloudinary": "^6.16.0",
-    "next-sanity-image": "^6.1.1",
     "next-seo": "^6.6.0",
     "node-fetch": "^2.6.7",
     "path-browserify": "^1.0.1",
