@@ -21,7 +21,7 @@ export async function POST(req: NextRequest) {
   const wh = new Webhook(signingSecret);
 
   // Get headers
-  const headerPayload = headers();
+  const headerPayload = await headers();
   const svix_id = headerPayload.get("svix-id");
   const svix_timestamp = headerPayload.get("svix-timestamp");
   const svix_signature = headerPayload.get("svix-signature");
