@@ -1,7 +1,11 @@
+/* istanbul ignore file */
 "use client";
 import { useRouter } from "next/navigation";
+import { useEffect } from "react";
 
 export default function NotFound() {
   const router = useRouter();
-  return router.replace("/404");
+  useEffect(() => {
+    router.replace("/404");
+  }, [router]);
 }
