@@ -1,30 +1,45 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { OakThemeProvider, oakDefaultTheme } from "@oaknational/oak-components";
 
-import { QuizResultQuestionStem } from "./QuizResultQuestionStem";
+import {
+  QuizResultQuestionStem,
+  type QuizQuestionStemProps,
+} from "./QuizResultQuestionStem";
 
+import type {
+  ImageItem,
+  TextItem,
+} from "@/node-lib/curriculum-api-2023/queries/pupilLesson/pupilLesson.schema";
 import { quizQuestions } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.new.fixture";
+import { MathJaxProvider } from "@/browser-lib/mathjax/MathJaxProvider";
 
 const meta: Meta<typeof QuizResultQuestionStem> = {
   component: QuizResultQuestionStem,
   decorators: [
     (Story) => (
-      <OakThemeProvider theme={oakDefaultTheme}>
-        <Story />
-      </OakThemeProvider>
+      <MathJaxProvider>
+        <OakThemeProvider theme={oakDefaultTheme}>
+          <Story />
+        </OakThemeProvider>
+      </MathJaxProvider>
     ),
   ],
 
   argTypes: {},
-} as Meta;
+};
 
 export default meta;
 
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<QuizQuestionStemProps>;
 
 const starterQuiz = quizQuestions;
 const mcqText = starterQuiz ? starterQuiz[0] : null;
 const mcqStemImage = starterQuiz ? starterQuiz[1] : null;
+
+// Define a sample question stem to ensure type correctness
+const sampleQuestionStem: (ImageItem | TextItem)[] = [
+  { type: "text", text: "This is a sample question text." },
+];
 
 /*
  * This is the view users will see on encountering an expired lesson
@@ -36,7 +51,7 @@ export const Default: Story = {
     return <QuizResultQuestionStem {...args} />;
   },
   args: {
-    questionStem: mcqText?.questionStem || [],
+    questionStem: sampleQuestionStem,
     displayIndex: 1,
   },
 };
@@ -46,7 +61,7 @@ export const Text: Story = {
     return <QuizResultQuestionStem {...args} />;
   },
   args: {
-    questionStem: mcqText?.questionStem || [],
+    questionStem: mcqText?.questionStem || sampleQuestionStem,
     displayIndex: 1,
   },
 };
@@ -56,7 +71,7 @@ export const Image: Story = {
     return <QuizResultQuestionStem {...args} />;
   },
   args: {
-    questionStem: mcqStemImage?.questionStem || [],
+    questionStem: mcqStemImage?.questionStem || sampleQuestionStem,
     displayIndex: 1,
   },
 };
